requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── LLM providers (install the one(s) you need) ────────────────────────────────
openai>=1.35.0
anthropic>=0.28.0

# ── Data / validation ──────────────────────────────────────────────────────────
pydantic>=2.7.0

# ── Environment & config ───────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── UI ─────────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Web research (company enrichment) ─────────────────────────────────────────
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
ddgs>=0.1.0

# ── Dev / testing ──────────────────────────────────────────────────────────────
pytest>=8.2.0
httpx>=0.27.0
